airflow/requirements.txt
openai==0.28.1
tqdm
pandas
python-dotenv
snowflake-connector-python
timedelta
dbt-snowflake


